cleanup: refactor BTF and use natively for context resolution
BTF has been available in all supported kernels for a long time, and
many code paths are likely broken without BTF. This change simply
requires BTF and documents how to provide external BTF if is it not
directly available.

This change is rather large because of the high level of interconnected
classes and components in the system. This is substantially reduced by
this change, and `BPFtrace&` is needed by a lot less internal plumbing.

Signed-off-by: Adin Scannell <amscanne@meta.com>

stack-info: PR: https://github.com/bpftrace/bpftrace/pull/4894, branch: user/amscanne/funcidx/3